Move MidiEvents to seperate file
